Update to build on top of DataTable commit
