+++ code/pluginbuild.xml
@@ -7350,6 +7350,13 @@ jar-all-plugins - Generate the plugin jar files
 	
 	<target name="jar-function-plugins" depends="makeplugindirs" description="Build (Link) Custom Function jar files">
 		<!-- PCGen custom functions -->
+		<jar jarfile="${lstplugins.dir}/Function-INPUT.jar" manifest="${src.java.dir}/plugin/lsttokens/manifest.mf">
+			<fileset dir="${build.classes.dir}">
+				<patternset>
+					<include name="plugin/function/InputFunction.class" />
+				</patternset>
+			</fileset>
+		</jar>
 		<jar jarfile="${lstplugins.dir}/Function-LOOKUP.jar" manifest="${src.java.dir}/plugin/lsttokens/manifest.mf">
 			<fileset dir="${build.classes.dir}">
 				<patternset>
