+++ code/src/java/plugin/function/InputFunction.java
@@ -15,7 +15,7 @@
  * along with this library; if not, write to the Free Software Foundation, Inc.,
  * 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
  */
-package pcgen.cdom.formula.function;
+package plugin.function;
 
 import java.util.Arrays;
